requirements.txt
beautifulsoup4
requests
PyYaml